requirements.txt
PyQt5
opencv-python-headless
pyzbar
requests
pyserial